requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
charset-normalizer==2.1.1
decorator==5.1.1
discord.py==2.1.0
executing==1.2.0
frozenlist==1.3.3
idna==3.4
ipython==8.6.0
jedi==0.18.1
matplotlib-inline==0.1.6
multidict==6.0.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
pure-eval==0.2.2
stack-data==0.6.0
traitlets==5.5.0
wcwidth==0.2.5
yarl==1.8.1
